changes response from decimal to float

+++ lib/aca_entities/fdsh/ssa/h3/contracts/requested_month_information_contract.rb
@@ -9,10 +9,10 @@ module AcaEntities
 
           params do
             required(:IncomeMonthYear).filled(:string)
-            required(:BenefitCreditedAmount).filled(:decimal)
-            optional(:OverpaymentDeductionAmount).maybe(:decimal)
-            optional(:PriorMonthAccrualAmount).maybe(:decimal)
-            optional(:ReturnedCheckAmount).maybe(:decimal)
+            required(:BenefitCreditedAmount).filled(:float)
+            optional(:OverpaymentDeductionAmount).maybe(:float)
+            optional(:PriorMonthAccrualAmount).maybe(:float)
+            optional(:ReturnedCheckAmount).maybe(:float)
             required(:PaymentInSuspenseIndicator).filled(:bool)
           end
         end
